+++ projects/ngx-scrollbar-demo/src/app/example-nested-virtual-scroll/card/card.component.ts
@@ -8,8 +8,7 @@ import { Component, ChangeDetectionStrategy, Input } from '@angular/core';
     </div>
   `,
   styleUrls: ['./card.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  standalone: true
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class CardComponent {