+++ projects/ngx-scrollbar-demo/src/app/example-scrollto-element/scroll-anchor.directive.ts
@@ -1,7 +1,6 @@
 import { Directive, Input, inject, ElementRef } from '@angular/core';
 
 @Directive({
-  standalone: true,
   selector: '[scrollAnchor]',
   host: {
     '[id]': 'id'