+++ projects/ngx-scrollbar/mat/src/public_api.ts
@@ -1,1 +1,2 @@
 export * from './mat-select-viewport';
+export * from './mat-select-scrollbar';