+++ projects/ngx-scrollbar/src/lib/thumb/thumb.ts
@@ -2,7 +2,6 @@ import { Directive } from '@angular/core';
 import { ThumbAdapter } from './thumb-adapter';
 
 @Directive({
-  standalone: true,
   selector: '[scrollbarThumbX]',
   providers: [{ provide: ThumbAdapter, useExisting: ThumbXDirective }]
 })
@@ -10,7 +9,6 @@ export class ThumbXDirective extends ThumbAdapter {
 }
 
 @Directive({
-  standalone: true,
   selector: '[scrollbarThumbY]',
   providers: [{ provide: ThumbAdapter, useExisting: ThumbYDirective }]
 })