+++ projects/ngx-scrollbar/src/lib/viewport/scroll-viewport.ts
@@ -1,7 +1,6 @@
 import { Directive, inject, ElementRef } from '@angular/core';
 
 @Directive({
-  standalone: true,
   selector: '[scrollViewport]'
 })
 export class ScrollViewport {